+++ src/main/java/com/duplex/house/repository/LogementRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.duplex.house.model.Logement;
 
-public interface LogementRepository extends JpaRepository<Logement, Long>{
+public interface LogementRepository extends JpaRepository<Logement, Integer>{
 
 }
